requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.1
langchain-community>=0.0.13
google-generativeai>=0.3.2
sqlalchemy>=2.0.25
pandas>=2.2.0
numpy>=1.26.0
streamlit>=1.29.0
plotly>=5.17.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pydantic>=2.5.2
aiohttp>=3.9.1
uvicorn>=0.25.0
fastapi>=0.108.0 